sched: Implement tapered dvfs_headroom
The linear one is very aggressive at the top frequencies. 25% headroom
means that we saturate the cluster very quickly. Given how expensive the
new clusters are at the top frequencies, it'd be smarter to taper the
headroom as we approach the max capacity of the cluster.

Implement that by simply taking the headroom of the different between
curret util and capacity of the CPU. The lower frequencies will get
a big headroom, but as we approach max capacity, we'll get almost no
headroom.

This behavior is control via procfs knob.

Only CFS + RT utilization use the tapered version. irq is usually small
and this can lead to potentially undesired boost almost when there are
no irqs running. DL shouldn't need any headroom in general. So keeping
it using the default non tapered version as well.

Bug: 261705301
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: I05bbe9a44a0fee48f3f7f11dadc691f039f1acda

[Reina: Port to the current state of sched]
Signed-off-by: M. "Harumajati" Alfarozi <reinazhard@gmail.com>