cpif: icmp packet type check for ping request
icmp packet type is checked for ping request for immididate delivery to modem which reduces ping req/resp latency.
rest of the packets type are delivered with 1 ms interval which reduces the number of irq to modem and
allow pcie link transition to L1.2 in case of no activity on pcie bus.

Change-Id: I02f846aace9e08a823d9f21132e79dbe99d44624
Signed-off-by: sharma.l@samsung.com