FROMGIT: workqueue: add cmdline parameter workqueue.panic_on_stall
When we want to debug the workqueue stall, we can immediately make
a panic to get the information we want.

In some systems, it may be necessary to quickly reboot the system to
escape from a workqueue lockup situation. In this case, we can control
the number of stall detections to generate panic.

workqueue.panic_on_stall sets the number times of the stall to trigger
panic. 0 disables the panic on stall.

Signed-off-by: Sangmoon Kim <sangmoon.kim@samsung.com>
Signed-off-by: Tejun Heo <tj@kernel.org>

Bug: 358922049
Change-Id: I727c011396f8bf956ca2184d28c43c30591ba6fe
(cherry picked from commit 073107b39e553a5ef911f019d4e433fd5a8601b7
https://git.kernel.org/pub/scm/linux/kernel/git/tj/wq.git for-next)
Signed-off-by: Sangmoon Kim <sangmoon.kim@samsung.com>