kcompressd: Increase nr_kcompressd
"The default value is actually designed to be the maximum number of
kcompressd that can run simultaneously. The current design is to create
the next kcompressd thread when all kfifos are full, and the number 4
is just a temporary setting.

You are right, changing it to match the number of CPUs might be better
to avoid confusion on how to set it up.

Best Regards,
Qun-wei"
- From LKML

Signed-off-by: M. "Harumajati" Alfarozi <reinazhard@gmail.com>